+++ consts.go
@@ -1,0 +1,6 @@
+package stocks
+
+const (
+	// YahooURL = Yahoo Finance
+	yahooURL = "https://query1.finance.yahoo.com/v7/finance/quote?symbols=%s"
+)