+++ queries.go
@@ -7,86 +7,11 @@ import (
 	"strconv"
 )
 
-const (
-	// YahooURL = Yahoo Finance
-	YahooURL = "https://query1.finance.yahoo.com/v7/finance/quote?symbols=%s"
-)
-
-/** TLR
- * * is the Top level json result
- */
-type TLR struct {
-	QuoteResponse QuoteResponse `json:"quoteResponse"`
-}
-
-/** Result
- * * full results of json response
- */
-type Result struct {
-	Symbol                            string  `json:"symbol"`
-	ShortName                         string  `json:"shortName"`
-	QuoteSourceName                   string  `json:"quoteSourceName"`
-	Language                          string  `json:"language"`
-	Region                            string  `json:"region"`
-	QuoteType                         string  `json:"quoteType"`
-	Triggerable                       bool    `json:"triggerable"`
-	Currency                          string  `json:"currency"`
-	Exchange                          string  `json:"exchange"`
-	MessageBoardID                    string  `json:"messageBoardId"`
-	ExchangeTimezoneName              string  `json:"exchangeTimezoneName"`
-	ExchangeTimezoneShortName         string  `json:"exchangeTimezoneShortName"`
-	GmtOffSetMilliseconds             int     `json:"gmtOffSetMilliseconds"`
-	Market                            string  `json:"market"`
-	EsgPopulated                      bool    `json:"esgPopulated"`
-	FirstTradeDateMilliseconds        int64   `json:"firstTradeDateMilliseconds"`
-	RegularMarketChange               float64 `json:"regularMarketChange"`
-	RegularMarketChangePercent        float64 `json:"regularMarketChangePercent"`
-	RegularMarketTime                 int     `json:"regularMarketTime"`
-	RegularMarketPrice                float64 `json:"regularMarketPrice"`
-	RegularMarketDayHigh              float64 `json:"regularMarketDayHigh"`
-	RegularMarketDayRange             string  `json:"regularMarketDayRange"`
-	RegularMarketDayLow               float64 `json:"regularMarketDayLow"`
-	RegularMarketVolume               int64   `json:"regularMarketVolume"`
-	RegularMarketPreviousClose        float64 `json:"regularMarketPreviousClose"`
-	FullExchangeName                  string  `json:"fullExchangeName"`
-	RegularMarketOpen                 float64 `json:"regularMarketOpen"`
-	AverageDailyVolume3Month          int64   `json:"averageDailyVolume3Month"`
-	AverageDailyVolume10Day           int64   `json:"averageDailyVolume10Day"`
-	StartDate                         int     `json:"startDate"`
-	CoinImageURL                      string  `json:"coinImageUrl"`
-	FiftyTwoWeekLowChange             float64 `json:"fiftyTwoWeekLowChange"`
-	FiftyTwoWeekLowChangePercent      float64 `json:"fiftyTwoWeekLowChangePercent"`
-	FiftyTwoWeekRange                 string  `json:"fiftyTwoWeekRange"`
-	FiftyTwoWeekHighChange            float64 `json:"fiftyTwoWeekHighChange"`
-	FiftyTwoWeekHighChangePercent     float64 `json:"fiftyTwoWeekHighChangePercent"`
-	FiftyTwoWeekLow                   float64 `json:"fiftyTwoWeekLow"`
-	FiftyTwoWeekHigh                  float64 `json:"fiftyTwoWeekHigh"`
-	FiftyDayAverage                   float64 `json:"fiftyDayAverage"`
-	FiftyDayAverageChange             float64 `json:"fiftyDayAverageChange"`
-	FiftyDayAverageChangePercent      float64 `json:"fiftyDayAverageChangePercent"`
-	TwoHundredDayAverage              float64 `json:"twoHundredDayAverage"`
-	TwoHundredDayAverageChange        float64 `json:"twoHundredDayAverageChange"`
-	TwoHundredDayAverageChangePercent float64 `json:"twoHundredDayAverageChangePercent"`
-	MarketCap                         int64   `json:"marketCap"`
-	SourceInterval                    int     `json:"sourceInterval"`
-	ExchangeDataDelayedBy             int     `json:"exchangeDataDelayedBy"`
-	Tradeable                         bool    `json:"tradeable"`
-	MarketState                       string  `json:"marketState"`
-}
-
-/** QuoteResponse
- * * is the top level json element from Yahoo Finance API
- */
-type QuoteResponse struct {
-	Result []Result    `json:"result"`
-	Error  interface{} `json:"error"`
-}
-
 /** GetDetails
  * * returns the base summary of a ticker from Yahoo Finance API
  */
 func GetFullDetails(ticker string) *Result {
-	res, err := http.Get(fmt.Sprintf(YahooURL, ticker))
+	res, err := http.Get(fmt.Sprintf(yahooURL, ticker))
 	if err != nil {
 		println(err)
 		return nil
@@ -114,7 +39,6 @@ func GetPriceAndPercentage(ticker string) (string, string, string) {
 
 	trimmedPrice := strconv.FormatFloat(det.RegularMarketPrice, 'f', 2, 64)
 	trimmedPercent := strconv.FormatFloat(det.RegularMarketChangePercent, 'f', 2, 64)
-	//trimmedPercent = trimmedPercent[1:]
 
 	var direction string = ""
 	if det.RegularMarketChangePercent < 0 {